+++ Modules/YugiTraderChatUI/Sources/Components/MessageBubble.swift
@@ -1,0 +1,21 @@
+//
+//  MessageBubble.swift
+//  YugiTraderChatUI
+//
+//  Created by Celan on 2023/06/05.
+//  Copyright © 2023 Celan. All rights reserved.
+//
+
+import SwiftUI
+
+struct MessageBubble: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct MessageBubble_Previews: PreviewProvider {
+    static var previews: some View {
+        MessageBubble()
+    }
+}
